leader_peer_handler now sends response back to leader_loop.
leader_loop has very primitive (and incompletely implemented) check for
majority commits and messages back to client when the majority is seen.
But only handles the happiest of happy paths properly.

Also test_leader_simple now hangs because it only starts a leader svr,
not any followers, so that test is commented out.