leader_loop now logs cmd to own log only after (majority-1) peers have logged it
This saves having to truncate leader logs on failed attempt to handle client cmds.
I think this is consistent with the raft spec, which is vague on the ordering here.
Also started to handle the "rejection scenario" where the AEReq is actively rejected
by a peer because it's log is out of date, but long ways still to go on log repair.
Cosmetic/comment changes to log.rs and peer.rs