rip everything apart; rebuild from the git history
This is a slightly finalized "massive restructure". The idea is to pull
the old code back out of the branches and drop it in where it applies,
rather than working side by side with semi-migrated code.

This should allow us to work faster, since we can skim the old codebase
looking for how to do stuff here.